Add ability to "capture" or get a screenshot of perf graphs
We often grab screenshots of perf graphs to quickly share with others or to use
in ad-hoc presentations. I think most people just used whatever screenshot
tool they had available or crop the image afterwards, which is annoying and if
you have a bunch of graphs you want to display next to each others it's really
hard to crop them all in the same place so they look decent.

This adds a button below each graph to capture a screenshot of the graph and
legend. Because of browser security issues, it's not possible to directly
generate a screenshot so we use html2canvas (http://html2canvas.hertzen.com/)
and generate an image from the canvas it produces.

From html2canvas' website:
"
This script allows you to take "screenshots" of webpages or parts of it,
directly on the users browser. The screenshot is based on the DOM and as such
may not be 100% accurate to the real representation as it does not make an
actual screenshot, but builds the screenshot based on the information available
on the page.
"

Instead of bundling a copy of html2canvas in our third-party directory, we use
cdnjs to grab a copy of html2canvas. There's lots of benefits to using a cdn
over bundling our own copy. I also updated jquery to use cdnjs instead of
jquery's cdn since cdnjs is crazy fast, and we only have to grab data from a
single cdn this way. It appears dygraphs is also available on cdnjs so I think
I'll update our scripts to use that and remove our bundled version from the
third-party dir in a separate commit.

Currently this opens up a new window and generates a static image that's the
same size as the original graph (and from there you can right click and save
the image or whatever.) It would be nice to avoid opening up another window
(since most browsers block this behavior by default until you tell it to allow
pop-ups from our site.) It would also be nice if you could specify the
width/height of the image you want to generate so you don't have to scale the
image after the fact.