Make FFTW self-validating
This commit makes FFTW self-validating in order to avoid
differences in precision that may occur across different
architectures or FFTW plans chosen for a given architecture.
It adds two config consts, one to say whether or not the
error values should be printed, and the other to define
the acceptable epsilon in error to tolerate.

Update the .goods to reflect the 'success' messages
expected.

While here, also did a "cute" cleanup to the repetition
in the 1D, 2D, 3D test configurations to loop over them
rather than list them explicitly which I noticed only
after code freeze.