Create pure python make_sys_basic_types.py.
It works exactly like other version, makeSysBasicTypes, but since it is pure
python it does not require "bc". "bc" is not a problem, to be clear, but many
distributions do not include it by default so it requires admin privileges to
install a package or the user has to install it from source. Python, on the
other hand, is more readily available in distros.

To verify it worked the same:

    ./util/config/make_sys_basic_types.py > b.chpl
    ./util/config/makeSysBasicTypes > a.chpl
    diff a.chpl b.chpl