Add formal types to Filerators
In chpldoc-ing the filerator routines, I realized that I'd not gotten
around to adding formal types and return types to them as I claimed I
preferred for standard library routines, but it was late enough in the
release cycle that I didn't want to do it then.  Making up for it now.

The one case that I didn't add a return type was the leader iterator
because (a) it seemed like it would only add fragility if I got it
wrong and (b) this interface isn't seen publicly by the user (at
least, presently).