Rework insertion of autoDestroys to use scope blocks rather than basic blocks.
The ad hoc algorithm for finding the last basic block in a scope did not always yield the
right answer.  We add autoDestroys for variables that are still owned at the end of their
declaration scopes, so it makes better sense to find the insertion point relative to the
scope block anyway.

Added a utility method insertAtExit() to insert statements in the flow where it exits a
given scoped block.  Renamed insertAtTailBeforeGoto as insertAtTailBeforeFlow, to more
accurately represent what it does.

Moved the isFlowStmt() predicate to expr.h, and removed the isCondStmt() conditional from
its body.  The isCondStmt() test was causing statements to be inserted in the wrong place
in some blocks.

Added insertAutoDestroyAtScopeExit().