Update test/man/ error when chpldoc or chpl does not exist.
Previously, if chpl or chpldoc did not exist, the test would run and then throw
an OSError when it tried to call the non-existent program with --help
flag. That is not very helpful.

Update test/man/ to check that the program exists, and if not reports a
specific error about the missing program.