Treat RVV as if global.
To ensure that the return-value variable is always owned on entry to the final block of a
function, treat it like a global variable: It is not tracked, and a bitwise copy into the
return=value variable is a consumer.

##DETAILS
###compiler/include/OwnershipFlowManager.h
Add a clause to isConsumed(), returning true if the LHS of a bitwise copy is the
return-value variable.

###compiler/util/OwnershipFlowManager.cpp
In extractSymbols(), do not add the return-value variable.
Add some cases in insertAutoCopy(), skipping autocopies in constructors, to avoid infinite
recursion.