Improve error messages for runaway comments
Previously, unterminated comments created errors like this:
 file.chpl:6: EOF in comment

I've updated the error message to be more useful with the start
of the comment and of the last nested comment:
 file.chpl:1: start of unterminated comment
 file.chpl:3: start of nested comment
 file.chpl:6: EOF in comment

To keep the PR diff small, I havn't committed the result of 'make parser'. I
will do so when merging, however.

test/parsing passes. There are no other tests that have the string "EOF in
comment".