Add a pass to remove ref-deref pairs introduced by cullOverReferences.
AMM needs to see when a temp is assigned to the return-value variable, because that
assignment is treated as a consumption.  If this assignment is piped through conversion to
a reference temp and then back to a value, ownership analysis misses that there is
effectively a bitwise assignment from the original temp to the RVV.

Down to 253 regressions (from 294).

##DETAILS
### b/compiler/include/OwnershipFlowManager.h
Enable debug printing #ifdef

### b/compiler/resolution/cullOverReferences.cpp
Factor cullOverReferences().
Add removeRefDerefPairs() to remove places where cullOverReferences causes the temp that
is eventually assigned to the RVV to go "underground".