Bug fix for extern block w/o clang-included runtime
The --llvm option to compileline changes CHPL_TARGET_PLATFORM in order to use
clang-included and prevent ABI issues when linking in the runtime. However,
when I originally added that flag I neglected to remember that runClang runs
both as an initial step in --llvm compiles but also as a parsing step with
extern { } blocks even without --llvm. When it's just parsing the extern
blocks, we shouldn't be changing the CHPL_TARGET_PLATFORM at all. Now we just
leave out  --llvm from the compileline flags if just_parse_filename is set.