Fix fileIO double-deletion.
In fileIO, and in several other tests using (you guessed it!) file I/O, we were deleting a
reference to the underlying string representation twice because the value returned from
modestring() was not getting properly copied but was still being treated (by AMM) as
owned.

I reinstated tracking of the RVV, and added the case where a bitwise copy from a global to
the RVV must be passed through an autoCopy call.

Down to 176 regressions (23 run-time).

##DETAILS
### b/compiler/include/OwnershipFlowManager.h
In bitwiseCopyArg(), removed the check that the RHS of a bitwise copy is also a local
variable.  This check must now be performed by client code.
In isConsumed(), added a case for the LHS being the RVV.

### b/compiler/util/OwnershipFlowManager.cpp
Re-enabled tracking of the RVV.
Moved processCreator() up.
Add logic for adding an autoCopy when the RHS of an assignment to the RVV is unowned.