Update selected tests or good files to make them pass.
##DETAILS

###spec/Records.tex
Revise added autoCopy function to just copy the value of C.
There was an extra "new" that was not correct (unless the return value is initialized
using "noinit").

###test/memory/shannon/freedMalloc.good
Line number 0 now prints as "--".
Updated behavior in mem tracking?

###test/types/string/StringImpl/memLeaks/find.good
An exact match with the base string should return "1", not "0".
(String indices are one-based.)