Adding a segfault for trying to assign a param member in a constructor
This test declares a class with a param member and tries to assign it
within the constructor (which is not legal).  Rather than generating a
useful error message, it segfaults.  Looking at the segfault, it
appears that a context that is expecting a direct variable reference
is actually getting an array access ('foo[0]'), yet it's not clear to
me why we would ever generate an array access to a param enum member
(or any param/enum for that matter).

The fix for the user is to avoid assigning to the param and rely on
Chapel's auto-transferring of param/type information from constructor
arguments to param/type fields of the same type.  But the user should
get a useful error message rather than a segfault.

This is something we should also work on cleaning up in our new
constructor world with its support for initialization sections
(specifically, it seems like assignments to param/type fields may be
something we'd consider legal in the initialization section since it's
constrained and similar to an initialization context).