Update genGraphs to work when PyYAML is available, but ANNOTATIONS file is missing.
Previously, genGraphs assumed an ANNOTATIONS.yaml file was available or
provided if it could import the annotate module, which only works when PyYAML
is installed. If PyYAML was installed, but an ANNOTATIONS.yaml file was not
found or specified, genGraphs ultimately tried open a NoneType file.

Update genGraphs to check for annotations file, while still checking for
ability to import annotate module.