Fix a few more memory leaks.
Too many autoDestroys were getting inserted because I forgot to clear out the OUT flow set
after using it to complete the USE set.

Also, we have to treat the value returned by _ddata.this(i: integral) as unowned.

Down to 211 run-time failures.

##DETAILS
###compiler/include/OwnershipFlowManager.h
Added code to obey the "return value is not owned" flag in isConsumed(SymExpr* se).

###compiler/util/OwnershipFlowManager.cpp
Added code to obey the "return value is not owned" flag in OwnershipFlowManager::computeTransitions().
We now clear the OUT sets at the top of OwnershipFlowManager::backwardFlowUse().
Added code to obey the "return value is not owned" flag in insertAutoCopy()