Specify cross-compilation for third-party packages on Cray-X* targets.
For third-party package builds on Cray X* target systems that use
autoconf for configuration, specify x86_64-unknown-linux-gnu as the
build system type (where we're building) and x86_64-cnl-linux-gnu as the
host system type (where the emitted code will run).  Making these
different tells autoconf (etc.) that we are cross-compiling, which among
other things prevents it from trying to run, on the build system, little
test executables compiled for the host system type.

The need for this was pointed out by our inability to build hwloc on a
Cray-XC system with the Intel compiler, optimizing for an ivybridge
target.  Not knowing we were cross-compiling led the configure script to
try to run an executable that had been built for ivybridge with -O3,
which led to an Intel library internal error message about an ISA
mismatch.  The message and an unexpected process exit status from the
executable then led to the configure script aborting.

Here we add the appropriate settings for gmp, hwloc, massivethreads,
qthreads, and tcmalloc.  For gmp and qthreads the logic builds on and
integrates with what was already there to handle target=knc.

While here, I did some other cleanup:
- gmp:
  - Allow for providing configure options from outside the Makefile by
    setting CHPL_GMP_MORE_CFG_OPTIONS.
  - Don't do the 'check' step if we're cross-compiling.
- massivethreads:
  - Put all the configure options in a CHPL_MASSIVETHREADS_CFG_OPTIONS,
    like we do for the other packages.
  - Allow for providing configure options from outside the Makefile by
    setting CHPL_MASSIVETHREADS_MORE_CFG_OPTIONS.
- tcmalloc:
  - Put all the configure options in a CHPL_TCMALLOC_CFG_OPTIONS, like
    we do for the other packages.
  - Allow for providing configure options from outside the Makefile by
    setting CHPL_TCMALLOC_MORE_CFG_OPTIONS.

The other third-party packages don't have this problem for one reason or
another; either they target the build system (for example llvm) or they
target the host system but they don't use autoconf (for example gasnet,
dlmalloc, re2).  These I didn't modify.