Rework forward flow, so OUT and EXIT are independent.
Down to 184 run-time errors.

##DETAILS
###compiler/include/bitVec.h
Add & and | dyadic operators.

###compiler/util/OwnershipFlowManager.cpp
Implemented the USED_LATER set.

In forwardFlowOwnership(), changed the forward flow equations so that OUT is not cancelled
by EXIT within the block where EXIT is true.  Rather, EXIT is used to prevent an OUT bit
from traversing the gap to succeeding blocks.
This simplifies the equation for determining where autoDestroys are needed: we just need
them where OUT and EXIT are both true.

Added checkForwardOwnership() to check that forward ownership is correct.  (Not yet enabled)

In insertAutoDestroys(), simplify the equation for finding where autoDestroy need to be
inserted to OUT & EXIT -- enabled by the above change to the forward flow equations.