Update .bads for five chpldoc futures
When we updated chpldoc testing to rely on .doc in the test name, we updated
all the .good files.  However, the .bad files for these futures were not
updated at the same time.  Testing missed this because the chpldoc directory is
getting skipped at the moment (investigating this issue as well).  There is one
more .bad failure whose cause I haven't determined yet.