Undo change to test/chpldoc/basics/main.doc.chpl
When the chpldoc tests were converted to .doc.chpl form, this test had its
.catfiles and .good updated to use main.doc instead of main.  However, since the
test declared its module name inside it, that update was not necessary.  When I
ran this future, it failed because of that update.

This was not detected in nightly testing because we have been accidentally
skipping chpldoc tests.  That should be alleviated by my previous commit.