Update docs() pass to check for directories before creating them.
Previously, when creating the directories for chpldoc output files, chpldoc
would iterator through each individual directory in the path, attempting to
create it. For example, if the output dir for a particular module was
/path/to/chapel/docs/modules/, it would try to create /, then /path, then
/path/to, etc. The mkdir() calls would fail, but the makeDir() helper only
generates a USR_FATAL() if the error it not EEXIST.

This worked ok and led to a simple solution... except on cygwin. On cygwin,
`mkdir /cygdrive/c` returns "Permission denied" error instead of an EEXIST.

Update the createDocsFileFolders() procedure, which is responsible for creating
all those directories, to only attempt creating directories that do not
exist. Add new existsAndDir() helper that uses stat() to determine if a
particular directory path exists and is a directory. Use new helper when
deciding whether to call makeDir().

Note that a path can exist, but not be a directory, and makeDir() will still be
called. This is expected and intentional, since makeDir() will report a useful
error message like, "Failed to create dir /blah due: <system error>".