Remove --compilelink support since it doesn't make sense now
The issue is that the Chapel runtime relies on symbols defined
in .chpl files, and so it doesn't make sense (in any simple way)
for C tests (or other C programs) to link with the Chapel runtime.

They can include the header files and then include use of the Chapel
runtime that will be sorted out when the Chapel compiler does the
linking.