Add iteratorInsertAutoDestroys.
Revised hoistRVVintoItsOwnScope to also hoist the return statement if the return value is
void or an immediate.

322 regressions (169 run-time)

##DETAILS
###compiler/AST/symbol.cpp
Revise FnSymbol::insertBeforeDownEndCount(Expr* ast) to accommodate my rework to place all
function bodies within an inner scope.

###compiler/include/OwnershipFlowManager.h
Add special code to insert autoDestroys earlier for iterators.
This is a preliminary version that places the autodestroy just after the last use.  I
think I want to use the zip1, zip2, ... concept to end pseudo-scopes at those boundaries.
I am not sure why this should work for a general iterator function, but it seems
attractive in comparison with my narrow-view implementation.  I have some debugging to do
yet, to see whether my approach flies or not.

###compiler/passes/insertAutoCopyAutoDestroy.cpp
Changed hoistRVVintoItsOwnScope to also hoist the return statement when the return type is
void or an immediate.
Moved computeTransitions after computExits, just in case it wanted to update exit points.
Added the call to iteratorInsertAutoDestroys().

###compiler/util/OwnershipFlowManager.cpp
Added the implementation for iteratorInsertAutoDestroys.