Fix chpldoc of enum, real, imag, and complex default values
Before, output for fields and globals of enums, reals, imags, and complexes
would display incorrect output if the field or global had a default value
associated with it.  In the case of reals, imags, and complexes, a default
value was detected but nothing after the " = " was output.  In the case of
enums, instead of

`enumName.enumConstant`

we would see

`.(enumName, "enumConstant")`

Both of these cases were solved by switching from prettyPrint to AstToText
for the type and init of a VarSymbol.  This was done by adding another public
method for AstToText, appendVarDef, which takes a VarSymbol and gathers
pertinent information from its defPoint.