Support annotations for xc single and multi-locale perf graphs
Previously the script that grabbed annotations computed hostname by effectively
doing `hostname` which worked for chap03, chap04, bradc-lnx because the graph
generation ran on the same machine as the testing.

We run xc multi-locale graph generation on a random linux box, and we don't use
the actual machine name it runs on anyways so we need some way to specify which
hostname to use rather than just using the real hostname of the machine the
script is running on.

Now instead of using `hostname` we use the 'name' that was passed to gengraphs.