See what happens if destructor stuff doesn't special-case sync/single
In a discussion about leaking syncs/singles today, I realized that
if we adopt the proposal to stop preserving variables beyond their
lexical scopes for the sake of potentially asynchronous tasks and
make this a user error instead, we can probably stop leaking syncs
and singles.

This is a five-minute fix (TM) attempt to treat them like other
variables and see what happens.