Updated per reviewer comments.
Renamed the gen/ directory as generate_tests/.

Removed the path from the filename included in a comment at the top of each test.

Added an entry for sin().

Did not add an entry for atan2(); it is dyadic so it needs a different template (beyond
the scope of this patch).

Did not add an entry for fabs(), since it is not supported in our Math module.  (It is
used as an extern support function, but is wrapped in the abs() public interface.)

Moved the auxilliary file monadic.txt into the test generation script as an initialized
variable (a list of tuples call "monadic").

Added README.md to describe how the auto test generation works.

Added the recommended path splicing, so the script can be run from any CWD.