Find unstructured exit points and put destructor code there as well.
Down to 16 run-time regressions.

##DETAILS

###compiler/include/stmt.h
###compiler/AST/stmt.cpp
Added BlockStmt::insertAtAllExitPoints().

###compiler/include/passes.h
###compiler/resolution/callDestructors.cpp
Made insertReferenceTemps() global.

###compiler/passes/insertAutoCopyAutoDestroy.cpp
Insert a global run of insertReferenceTemps when done.

###compiler/util/OwnershipFlowManager.cpp
Now insert autoDestroys at all exits from a scope block.

###modules/internal/String.chpl
Updated this() to also reject negative indices.

###test/types/string/StringImpl/memLeaks/substring.future
Converted this test to a future because there is a mismatch beween the test output and the
.good file.  Whether substring should halt on an invalid index or just return an empty
string is a design decision beyond the scope of this patch.