Rework gamma function promotion tests to avoid a nan vs. inf ambiguity.
The gamma function passes through infinity at zero and negative integer input values.  An
exact representation of the input should give "nan" as the output because at those
singularities, it is ambiguous whether the value of the function is -inf or +inf.  At a
slight deviation from those exact input values, however, the function will resolve to a
very large positive or negative number.  In that case the floating-point representation
will be able to express it as a representable value, or as either "inf" or "-inf".

To avoid the ambiguity in the output, I changed the discrete domain over which the gamma
functions are computed, so it excludes negative integers and zero.  The domain change was
made in make_promotion_tests.py which was re-run to produce new .chpl files.  Then the
resulting output was used to update the .good files.