If chpldoc is not present, skip all remaining chpldocopts for chpldoc test
It was observed that test/chpldoc/compflags/folder/failToCreateDir would fail
when the tester did not have chpldoc built in their environment.  This test had
two compiler options in its .chpldocopts file, but the check to warn and skip
testing if chpldoc was not built only exited the current compopt line, not all
compopt lines.  In that section, the "compiler" value was changed, and so the
second option would not enter the same if statement to determine that it, too,
needed to be skipped.  This changes the behavior of that section so that all
compopt lines were skipped when chpldoc was detected as unbuilt, instead of just
the first.