Don't speculatively build with the PGI compiler on Cray X*.
We have problems involving alloca() in gmp when it's built with the PGI
target compiler on Cray X* platforms.  So, don't speculatively build gmp
in that situation.  The specific file changes are as follows:

third-party/Makefile:
Set a GMP_SPECULATIVE make variable when invoking the third-party/gmp
Makefile for a speculative build.

third-party/gmp/Makefile:
For speculative builds, squash them entirely and print a message when
target=cray-x* and target compiler=*pgi.  For non-speculative builds
with this target and target compiler, warn that the resulting gmp may
have problems, but do the build anyway.

While here, I also rearranged the make targets so that we don't unpack
the tar file or do the configure step unless we're going to build gmp.
With the previous arrangement, there wasn't a good place to squash the
build until after the unpack and configure had been done, which take a
bit of time.

While here, for a 'clean', delete not only the configuration-specific
build subdir but also the config-specific install subdir.  This makes
things a little easier when testing Makefile changes, since for the
speculatively built packages we depend not only on the build subdir but
also the install subdir (for the .h file, whose presence tells us that
we have already speculatively built).