Remove field-wise initialization in default constructor wrapper for non-tuples.
Field-wise assignment in the default-constructor wrapper (for non-tuple types) was causing
autoCopy to be called redundantly on record argument.  Also, it's completely redundant
with the field-wise assignment in the compiler-generated field-wise constructor anyway, so
I pulled it out.

A few tests had to be tweaked to avoid use-after-free errors, but this mostly passed
(Yippee!).

##DETAILS
###modules/internal/ChapelArray.chpl
Test userAssoc2 was failing due to use-after-free.  Distributions are in general classes,
so calls to their copy constructors are moot.  The constructor for a distribution must
increment the reference count after all.
(This change probably also needs to be propagated to _getDistribution, and to those
routines for arrays and domains as well.  We need to look elsewhere for the reason that
array reference counts are not going back to zero.  However, the main change in this
commit surely helps.)

###modules/standard/List.chpl
Give Lists copy semantics.  Our other alternatives are to always leak them, or to perform
some kind of reference counting.

###test/classes/bradc/records/assignRecord2.chpl
Explicit destroys are no longer needed, since AMM takes care of this.
The test should still work if they are left in, but didn't for some reason, so I took them
out.  Probably the failure has to do with destroying a clone rather than the original (and
then accessing the list through the original).

###test/types/seq/bradc/returnSeq2.good
###test/types/seq/deitz/test_seq4.good
The semantics of list copying has changed from reference-like to value-like.
As suggested above, reference-like semantics either require reference counting or force us
to leak memory at program exit.