Pull test_module_Sort-baseline.
This test was healed by https://github.com/chapel-lang/chapel/commit/665dd04.

The change flattens the body of a function into the block that contains the scope of the
return value variable.  Then, because this caused a jump to the return label to fail to be
considered a jump that exited that inner block (the one that no longer exists), a flag was
added to the goto statement to indicate if it jumps to the return label.  In code that
inserts autoDestroys, this kind of goto is treated specially, so it still appears to cause
flow to exit the body of the function.

Before the change, this test case was encountering too many autoDestroy calls and now it
doesn't.  The original problem was related to iterator lowering.  So (without comparing
old and new code side-by-side) we can guess that simplification of the AST allows iterator
lowering to encounter autodestroys in the right place now, whereas before it did not.