Prevent autoCopy from being applied to any argument to an autoDestroy call.
There was already code in place to prevent the addition of an autoCopy to the actual
passed to an autoDestroy inside of a destructor or autoDestroy function.  The predicate
has been broadened so it now returns true for the argument of an autoDestroy call
regardless of its context.

The test case that was failing (among others) was simple-record.  This was due to the
record destructor attempting to increment the ref count of a distribution that had already
been deleted.  Removing the call to the record *con* structor before the call to the
record destructor in _ddata_free resolved this problem.

Down to 209 regressions > 54 run-time > 27 ferguson