Remove redundant autocopy of return value variable.
There were extra autocopies being inserted in this functions, due to the RHS of a copy
into the RVV being untracked.  This was corrected by adding a clause to insertAutoCopy()
to avoid that insertion for the RVV in a return primitive.  This means that there must
always be a RVV and it must gain ownership of its contents by the time control reaches the
return statement.

This change made the exclusion of autocopies in chpl__ensureDomainExpr incorrect, so that
special-purpose code was removed from isRetVarCopyInConstructor.