Remove a couple of undesired comments and fix some white space.
Comments naming the specific symbols a #include is intended to supply
are hard to maintain and, as a result, often wrong.  Remove two such
comments, at least one of which has become incorrect since it was added.

This is the first in possibly a series of commits to clean up the
Qthreads tasking shim before pushing our changes back to the Qthreads
repo.