Allow for errors when touching the *.m4 files.
If we're working with a non-Chapel Qthreads repo, some or all of that
repo's *.m4 files may be symbolic links to corresponding system files
found by autogen.sh.  If so, their permissions won't allow us to touch
them in order to change their timestamps.  Not changing the timestamps
won't cause a problem, but the touch failure halts our make.  Arrange
for it not to do so.  (One can use a non-Chapel Qthreads repo by setting
QTHREAD_SUBDIR on the make command, to point to its directory.  This is
most often done by Qthreads developers wanting to try out Qthreads
changes made in their own repos rather than in the Chapel copy.)