Updates to formatted I/O
Remove %# in format strings
Update integer reading and writing to use precision in addition to
showpoint. That way, writef("%.2n", 1) and writef("%{#.##}") will
both print print 1.00. Also allows #### to start with a period as in
writef("%{.##}", 1) when all that is desired is to specify the precision.

Update _peek_number_unlocked to support allow_point instead of
just allow_real in order to handle these cases with integers.