Rename memory diagnostics symbols to clarify their purpose.
The config consts, module procs, and runtime functions associated with
memory diagnostics had symbol names that were unclear in some cases and
downright misleading in others.  Here we rename things to clarify their
purposes and relationships.

The old situation was as follows.  Each table entry shows a kind of
memory allocation report along with the config const that enabled doing
that report automatically at normal termination, the Standard/Memory
module proc that was called to produce that report, and the runtime
function which actually printed the report.

  Functionality			Config
  -------------			---
  statistics			config:  memStats
      	    			proc:    printMemStat
      	    			runtime: chpl_printMemStat
  summary by allocation type	config:  memLeaks
      	    			proc:    printMemLeaksTable
      	    			runtime: chpl_printLeakedMemTable
  details of all allocations	config:  memLeaksTable
      	    			proc:    printMemTable
      	    			runtime: chpl_printMemTable

The new situation is as follows.  Only the memStats config const has not
been renamed.

  Functionality			Config
  -------------			---
  statistics			config:  memStats
      	    			proc:    printMemAllocStats
      	    			runtime: chpl_printMemAllocStats
  summary by allocation type	config:  memLeaksByType
      	    			proc:    printMemAllocsByType
      	    			runtime: chpl_printMemAllocsByType
  details of all allocations	config:  memLeaks
      	    			proc:    printMemAllocs
      	    			runtime: chpl_printMemAllocs

Beyond the renamings there are some additional changes to a few files,
as follows.

modules/standard/Memory.chpl:
The old module procs remain, but now they just print a "deprecated"
message and invoke the new ones.

runtime/src/chplmemtrack.c:
Functions that used to invoke error() and halt when called without
--memTrack being thrown (perhaps implicitly) at execution time now
invoke warning() and return instead.  It seemed overly harsh to kill off
the entire program just because the memory diagnostics could not be
produced.
The by-type memory allocation report's headers used to refer to "leaked"
memory.  Strictly speaking that wasn't accurate, since allocated memory
only represents a leak if it is being reported at termination due to one
of the config consts having been set.  Now the headers only refer to
"leaked" memory when the report is being produced at termination time.
For an explicit invocation during execution, they refer to "allocated"
memory.
While I was here I also moved the chpl_reportMemInfo() function so that
it follows the others, since that seemed a more logical place for it.

doc/release/README.executing:
(This should have been done as part of the chpldocing work, just prior
to the 1.11 release, but I missed it.)  Remove the discussion of the
memory diagnostics related config consts, and replace it with a pointer
to the corresponding chpldoc information and a brief symopsis telling
what each config const does.