Add new implementation of collectLiveLocalVariabels.
The existing version only test for liveness of a local variable crossing the beginning or
end of a loop if the structure is a single-loop iterator.  But the problem (where a
local variable is uninitialized on second and subsequent calls to an iterator advance
function) exist in all loop structures.

I generalized the routine to use basic blocks to determine where loops exist (by looking
for back edges), so that local variables crossing the beginning or end of *any* loop are
converted into fields in the iterator class.

The implementation is slightly aggressive, because I don't check that a loop actually
contains a yield.

Now at 222 regressions > 45 runtime > 11 non-leak.