Rework insertLocalsForRefs() to handle the no def case.
AMM moved the scope of a ref temp to be equal to the scope of the variable it references.
In some cases iterator lowering results in the definition of a reference temp (i.e. what
it pointed to) to be lost.

I don't completely understand the mechanism behind this, but created a workaround: If the
ref temp is undefined in the calling context, I insert a temporary variable and make the
ref point to that.  This insertion is repeated at every use of that ref temp.

It would probably be better to identify this case up front, and move the variable into the
iterator as a local variable.  The case where I see it appear is in a coforall statement
that is inlined.

Now at 201 regressions > 46 run-time > 11 non-leak.