Skip gdbSetConfig.chpl on the whitebox testing for XC
We had been experiencing sporadic hangs with this test, which consists of a
single writeln of a config variable and then runs with the --gdb flag during
whitebox testing of our XC TARGET builds (and infrequently our TARGET-noPrgEnv
builds).  The hang occurred during program shutdown once the test program had
actually completed, and disappeared when the tasking layer was switched to
fifo or when we switched to the same builds aimed at an XE.  The problem also
disappeared when run on a proper XC machine.  With this information in mind,
it seems likely that the problem is specific to our whitebox installation and
not an actual issue - so the solution is to skip the test for these
configurations.