Convert dos-formatted tests to unix
This is a trivial change in which I ran dos2unix on tests that
seemed clearly in DOS mode w.r.t. CRLFs.  Typically, I've done
this upon encountering such cases, but this can cause issues
for longstanding changes in terms of merge conflicts (one of
which just bit me).