Provide for not tracking certain internal memory descriptor types.
Add a "track this" indicator to the table of memory descriptors, along
with a function to return that indicator and code to pay attention to it
in the memory tracking implementation.  This will allow us to not track
certain runtime internal allocations that don't have a well-defined
connection to constructs in Chapel code, such as internal descriptors
for Active Memory handling.

For now all such indicators are actually set so that everything is still
tracked.  In the future I will change some of them to turn tracking off,
in order to help resolve https://chapel.atlassian.net/browse/CHAPEL-8.

The modification with the largest impact in this commit is actually not
a functional one.  In order to reformat the CHPL_MD_ALL_MEMDESCS() macro
definition to make it still be readable with the addition of the "track"
indicator, I shortened a number of the memory descriptor enum names and
some of the corresponding strings.  This part of the commit is wholly
responsible for the following files changing:
  runtime/src/chpl-privatization.c
  runtime/src/chpl-string-support.c
  runtime/src/chpl-string.c
  runtime/src/comm/gasnet/comm-gasnet.c
  runtime/src/comm/none/comm-none.c
  runtime/src/config.c
  runtime/src/tasks/fifo/tasks-fifo.c
  runtime/src/threads/pthreads/threads-pthreads.c

The other files changed as follows.

runtime/include/chpl-mem-desc.h
runtime/src/chpl-mem-desc.c
  Add the "track" indicator to CHPL_MD_ALL_MEMDESCS() and adjust uses of
  this macro to take that indicator into account.  Make some associated
  internal changes to improve readability.  Add chpl_mem_descTrack(), to
  return the "track" indicator.
  And of course, while here, change a number of the enum names and a few
  of the corresponding description strings, as described above.

runtime/src/chplmemtrack.c
  Track or don't track specific memory descriptor types depending on the
  setting of the "track" indicator for that type and the value of the
  overriding CHPL_RT_MEMTRACK_ALL_MDS environment variable.

runtime/include/chplsys.h
runtime/src/chplsys.c
  Add chpl_env_getBool(), which gets the presumed boolean value of a
  specified CHPL_RT_* environment variable, with a default.  (This new
  function can also now be reused in a few other places in the runtime
  where we need the value of a boolean environment variable.)